+++ src/lib/components/__tests__/OverlayView.test.tsx
@@ -1,33 +1,25 @@
 import React from 'react'
 import 'jest-dom/extend-expect'
-import loadjs from 'loadjs'
 import 'react-testing-library/cleanup-after-each'
-import {render, wait, cleanup} from 'react-testing-library'
+import {render, cleanup, flushEffects} from 'react-testing-library'
 import {OverlayView, MapBox} from '../..'
 import {GoogleMapProvider} from '../../contexts/GoogleMapContext'
 import {defineGlobalVariable} from '../../__test__helpers__'
 
-describe('OverlayView', () => {
-  beforeEach(() => {
-    defineGlobalVariable()
-    jest.spyOn(loadjs, 'reset')
-  })
+defineGlobalVariable()
 
+describe('OverlayView', () => {
   afterEach(() => {
     cleanup()
-    Object.defineProperty(global, 'google', {value: undefined})
-    jest.restoreAllMocks()
   })
 
   it('renders inside a MapBox', async () => {
-    const {container} = render(
+    render(
       <GoogleMapProvider>
-        <MapBox apiKey="A_FAKE_API_KEY" opts={{}} />
+        <MapBox apiKey="A_FAKE_API_KEY" />
         <OverlayView />
       </GoogleMapProvider>,
     )
-    await wait(() => {
-      expect(container.innerHTML).not.toMatch('Loading...')
-    })
+    flushEffects()
   })
 })